# longaeva/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ActionRowRead": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "enum": [
              "hold",
              "add",
              "trim",
              "exit"
            ],
            "title": "Action",
            "type": "string"
          },
          "costs": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CostComponentsRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "gaps": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ValueGapRead"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gaps"
          },
          "is_no_action": {
            "title": "Is No Action",
            "type": "boolean"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "selected": {
            "title": "Selected",
            "type": "boolean"
          },
          "target_shares": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Shares"
          }
        },
        "required": [
          "action",
          "label",
          "is_no_action",
          "selected",
          "target_shares"
        ],
        "title": "ActionRowRead",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MultipleRangeInput": {
        "additionalProperties": false,
        "description": "Caller-supplied earnings multiple. Order is checked by the bridge, not here.",
        "properties": {
          "high": {
            "title": "High",
            "type": "number"
          },
          "low": {
            "title": "Low",
            "type": "number"
          },
          "mid": {
            "title": "Mid",
            "type": "number"
          }
        },
        "required": [
          "low",
          "mid",
          "high"
        ],
        "title": "MultipleRangeInput",
        "type": "object"
      },
...
      "PositionRead": {
        "additionalProperties": false,
        "properties": {
          "direction": {
            "const": "long",
            "title": "Direction",
            "type": "string"
          },
          "shares": {
            "title": "Shares",
            "type": "number"
          },
          "source": {
            "enum": [
              "decision_rule",
              "request_override"
            ],
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "shares",
          "direction",
          "source"
        ],
        "title": "PositionRead",
        "type": "object"
      },
...
      "ReferencePriceRead": {
        "additionalProperties": false,
        "properties": {
          "acceptance_utc": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acceptance Utc"
          },
          "period_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period Label"
          },
          "price": {
            "title": "Price",
            "type": "number"
          },
          "price_quote": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Quote"
          },
          "price_source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Source Id"
          },
          "source": {
            "enum": [
              "sec_buyback_average",
              "request_override"
            ],
            "title": "Source",
            "type": "string"
          },
          "source_label": {
            "title": "Source Label",
            "type": "string"
          }
        },
        "required": [
          "price",
          "period_label",
          "acceptance_utc",
          "source",
          "source_label",
          "price_source_id",
          "price_quote"
        ],
        "title": "ReferencePriceRead",
        "type": "object"
      },
...
      "ValuationActionsRead": {
        "additionalProperties": false,
        "properties": {
          "actions": {
            "items": {
              "$ref": "#/components/schemas/ActionRowRead"
            },
            "title": "Actions",
            "type": "array"
          },
          "bridge": {
            "$ref": "#/components/schemas/ValuationBridgeRead"
          },
          "decision": {
            "enum": [
              "hold",
              "add",
              "trim",
              "exit"
            ],
            "title": "Decision",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "margin": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Margin"
          },
          "position": {
            "$ref": "#/components/schemas/PositionRead"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "reference_price": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReferencePriceRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "rule": {
            "$ref": "#/components/schemas/DecisionRuleRead"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "ok",
              "unsupported"
            ],
            "title": "Status",
            "type": "string"
          },
          "value_per_share": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value Per Share"
          }
        },
        "required": [
          "status",
          "reason",
          "label",
          "run_id",
          "bridge",
          "reference_price",
          "position",
          "rule",
          "decision",
          "margin",
          "value_per_share",
          "actions"
        ],
        "title": "ValuationActionsRead",
        "type": "object"
      },
      "ValuationActionsRequest": {
        "additionalProperties": false,
        "description": "Illustrative actions for one saved run. The decision rule is not editable here.",
        "properties": {
          "diluted_shares": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diluted Shares"
          },
          "multiple_range": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MultipleRangeInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "net_interest_other": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Interest Other"
          },
          "reference_price": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Price"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "shares": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shares"
          },
          "tax_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tax Rate"
          }
        },
        "required": [
          "run_id"
        ],
        "title": "ValuationActionsRequest",
        "type": "object"
      },
      "ValuationBridgeRead": {
        "additionalProperties": false,
        "properties": {
          "assumptions": {
            "items": {
              "$ref": "#/components/schemas/AssumptionRead"
            },
            "title": "Assumptions",
            "type": "array"
          },
          "cutoff_ts": {
            "format": "date-time",
            "title": "Cutoff Ts",
            "type": "string"
          },
          "earnings_driven": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EarningsDrivenRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "forward_eps": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ForwardEpsRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "grid": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ValueCellRead"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grid"
          },
          "metric": {
            "title": "Metric",
            "type": "string"
          },
          "multiple_band": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MultipleBandRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "multiple_driven": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MultipleDrivenRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "n_paths": {
            "title": "N Paths",
            "type": "integer"
          },
          "n_quarters": {
            "title": "N Quarters",
            "type": "integer"
          },
          "outer_envelope": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OuterEnvelopeRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "ok",
              "unsupported"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "reason",
          "run_id",
          "cutoff_ts",
          "metric",
          "n_paths",
          "n_quarters",
          "assumptions"
        ],
        "title": "ValuationBridgeRead",
        "type": "object"
      },
      "ValuationBridgeRequest": {
        "additionalProperties": false,
        "properties": {
          "diluted_shares": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diluted Shares"
          },
          "multiple_range": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MultipleRangeInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "net_interest_other": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Interest Other"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "tax_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tax Rate"
          }
        },
        "required": [
          "run_id"
        ],
        "title": "ValuationBridgeRequest",
        "type": "object"
      },
      "ValuationMultiplesRead": {
        "additionalProperties": false,
        "properties": {
          "band": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MultipleBandRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "cutoff_ts": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cutoff Ts"
          },
          "method_label": {
            "title": "Method Label",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/PeHistoryRowRead"
            },
            "title": "Rows",
            "type": "array"
          },
          "source_basis": {
            "title": "Source Basis",
            "type": "string"
          },
          "status": {
            "enum": [
              "ok",
              "unsupported"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "reason",
          "cutoff_ts",
          "source_basis",
          "method_label",
          "rows"
        ],
        "title": "ValuationMultiplesRead",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Standalone Visa business-simulation API (hackathon).",
    "title": "Longaeva",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/valuation/actions": {
      "post": {
        "operationId": "post_actions_valuation_actions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValuationActionsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValuationActionsRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Actions",
        "tags": [
          "valuation"
        ]
      }
    },
    "/valuation/bridge": {
      "post": {
        "operationId": "post_bridge_valuation_bridge_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValuationBridgeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValuationBridgeRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Bridge",
        "tags": [
          "valuation"
        ]
      }
    },
...
    "/valuation/multiples": {
      "get": {
        "operationId": "get_multiples_valuation_multiples_get",
        "parameters": [
          {
            "in": "query",
            "name": "cutoff_ts",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cutoff Ts"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValuationMultiplesRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Multiples",
        "tags": [
          "valuation"
        ]
      }
    },
...
  }
}